divided into the folder. updated the namespace's name
