+++ Infastructure/Services/FridgeModelService.cs
@@ -5,7 +5,7 @@ using AutoMapper;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Infastructure.Persistence.Services
+namespace Infastructure.Services
 {
     public class FridgeModelService : IFridgeModelService
     {
